Separate cross margin account overview and settings queries
